🐛 fix(logger): standardize log datetime format
This commit improves the `insertNewLog` function in the `logger.js` file.
The function is responsible for inserting a new log entry into the "Logs"
sheet of the bound Apps Script. The log entry includes the datetime,
username, URL, filetype, and filename of the download activity.

The following changes were made:
- The datetime parameter now accepts a JavaScript Date object instead of
  a string text
- Additional lines were added to retrieve the time zone from the active
  spreadsheet
- The value of the `datetime` parameter is converted to a datetime string
  in the format of "yyyy-MM-dd HH:mm:ss" using the `Utilities.formatDate()`
  method

This commit also modifies the `fetcher.js` file. The change makes the
`getFileDetails` function pass the date and time as a Date object when
calling the `insertNewLog` function, removing the `toLocaleString()`
appended to the `datetime` variable.

Note: The use of `Utilities.formatDate()` instead of `toLocaleString()`
      provides better performance for date formatting.

Fixes #77 and #80